requirements.txt
# 核心Python依赖（可直接用于：pip install -r requirements.txt）
aiofiles>=24.1.0
fastapi>=0.116.1
folium>=0.14.0
geopandas>=1.1.1
geopy>=2.4.1
httpx>=0.28.1
langchain>=0.3.26
langchain-deepseek>=0.1.3
langchain-mcp-adapters>=0.1.9
langchain-tavily>=0.2.11
langgraph-cli[inmem]>=0.3.4
matplotlib>=3.10.3
mcp>=1.11.0
numpy>=1.24.0
openai>=1.95.1
pandas>=2.0.0
pydantic>=2.11.7
pyproj>=3.6.0
python-dotenv>=1.1.1
redis>=6.2.0
requests>=2.28.0
scikit-learn>=1.7.0
shapely>=2.0.0

# 注意：
# - SUMO 的 traci/sumolib 通常随 SUMO 安装提供，不通过 pip 安装
# - 如需强化学习/优化能力，通常还需安装 PyTorch（torch/torchvision/torchaudio），建议参考 PyTorch 官方安装方式


